[TS-33] Refactor: component name change, footer btn add
