+++ src/pages/WithdrawalPage.tsx
@@ -1,5 +1,5 @@
-import CommonDelete from "@/components/common/CommonDelete/CommonDelete";
 import Header from "@/components/common/Header/Header";
+import SelectReason from "@/components/common/SelectReason/SelectReason";
 
 import { modalType } from "@/constants/modalConstants";
 import { leaveReasonData } from "@/constants/myPageConstants";
@@ -10,7 +10,7 @@ const WithdrawalPage = () => {
 			<Header>
 				<Header.PrevButton />
 			</Header>
-			<CommonDelete
+			<SelectReason
 				title="어떤 이유로 탈퇴하시나요?"
 				reasonDefaultText="탈퇴 이유를 선택해 주세요"
 				modalType={modalType.LEAVE_REASON}
